requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
fastapi==0.115.6
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-geminidataanalytics==0.2.0
googleapis-common-protos==1.69.1
httpx==0.28.1
httpx-oauth==0.16.1
jsonschema==4.23.0
numpy==2.2.3
pandas==2.2.3
proto-plus==1.26.0
protobuf==5.29.3
pydantic==2.10.6
python-dotenv==1.0.1
python-multipart==0.0.20
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn[standard]==0.32.1
